requirements.txt
Django==4.2.3
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
djoser==2.2.0
pipenv==2023.9.8
virtualenv==20.24.3
django-extensions==3.2.3
django-filter==23.2
